requirements.txt
# requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
sqlalchemy==2.0.19
psycopg2-binary==2.9.6
pydantic==2.0.3
python-dotenv==1.0.0
unicodedata2==15.0.0
pytest==7.4.0
black==23.7.0
pylint==2.17.5